[Remove, Rename] #15 테스트 코드 디렉토리 정리
